api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
celery[redis]==5.3.6
redis==5.0.1
yt-dlp==2025.12.8
openai-whisper==20231117
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
